Move filter testing into per-language files  * Make test-filter.js do generic, language-agnostic tests  * Move each language into its own test file that only tests that language's filter. Allows for more targeted testing and more reliability that we're testing what we think we are.
