+++ lib/filter.js
@@ -1,22 +1,36 @@
 var fs = require('fs');
 var path = require('path');
 
-var TweetFilter = function(filter_dir, excluded_terms) {
-    this.filters = this.loadFilters(filter_dir);
+var TweetFilter = function(filters_or_dirpath, excluded_terms) {
+    if (typeof filters_or_dirpath === "string") {
+        this.filters = TweetFilter.getFiltersFromDirectory(filters_or_dirpath);
+    } else {
+        this.filters = filters_or_dirpath;
+    }
     this.excluded_terms = excluded_terms || [];
 }
 
 // load all files from a directory and use their exports as filters
 // the files should export arrays of strings/regex expressions
-TweetFilter.prototype.loadFilters = function(dir) {
+TweetFilter.getFiltersFromDirectory = function(dir) {
     dir = path.resolve(dir);
-    var filters = {};
     var filter_files = fs.readdirSync(dir);
-    filter_files.forEach(function(filter_file) {
+    var joined_files = [];
+    filter_files.forEach(function(file) {
+        joined_files.push(path.join(dir, file));
+    })
+    return TweetFilter.getFiltersFromFiles.apply(this, joined_files);
+}
+
+// load the file(s) and use its/their exports as filters
+// the file should export arrays of strings/regex expressions
+TweetFilter.getFiltersFromFiles = function() {
+    var filters = {};
+    for (var i = 0; i < arguments.length; i++) {
+        var filter_file = path.resolve(arguments[i]);
         var filter_name = path.basename(filter_file, path.extname(filter_file));
-        var filter = require(path.join(dir, filter_file));
-        filters[filter_name] = filter;
-    });
+        filters[filter_name] = require(filter_file);
+    }
     return filters;
 }
