+++ test/test-spanish.js
@@ -1,0 +1,22 @@
+var shared = require('./lib/shared');
+var TweetFilter = require('../lib/filter')
+var filter = new TweetFilter(TweetFilter.getFiltersFromFiles("filters/spanish"))
+
+var matches = [
+    "hi i am spanish and Quisiera Ser Vegana",
+    "quisiera hacerme vegana they tell me",
+    "i also think that quiero hacerme vegana is something else I say",
+];
+
+var falsePositives = [
+];
+
+exports.matches = function(test) {
+    shared.testMatches(test, matches, filter);
+    test.done();
+}
+
+exports.falsePositives = function(test) {
+    shared.testFalsePositives(test, falsePositives, filter);
+    test.done();
+}
